refactor repository module (#2514)
* refactor repository module

* remove duplicate code

* fix table names
