+++ src/modules/repository/implementation/sequelize/repositories/operation-repository.js
@@ -1,0 +1,35 @@
+class OperationRepository {
+    constructor(models) {
+        this.sequelize = models.sequelize;
+        this.models = { get: models.get, publish: models.publish, update: models.update };
+    }
+
+    async createOperationRecord(operation, operationId, status) {
+        return this.models[operation].create({
+            operation_id: operationId,
+            status,
+        });
+    }
+
+    async getOperationStatus(operation, operationId) {
+        return this.models[operation].findOne({
+            attributes: ['status'],
+            where: {
+                operation_id: operationId,
+            },
+        });
+    }
+
+    async updateOperationStatus(operation, operationId, status) {
+        await this.models[operation].update(
+            { status },
+            {
+                where: {
+                    operation_id: operationId,
+                },
+            },
+        );
+    }
+}
+
+export default OperationRepository;
