+++ src/modules/repository/implementation/sequelize/repositories/user-repository.js
@@ -1,0 +1,16 @@
+class UserRepository {
+    constructor(models) {
+        this.sequelize = models.sequelize;
+        this.model = models.user;
+    }
+
+    async getUser(username) {
+        return this.model.findOne({
+            where: {
+                name: username,
+            },
+        });
+    }
+}
+
+export default UserRepository;
